server/requirements.txt
fastapi[all]
uvicorn==0.30.1
python-multipart==0.0.9
Pillow
pytest==8.2.2
httpx==0.27.0
langchain-core==0.3.27
langchain-google-genai==2.0.8
langgraph==0.2.34
python-dotenv==1.0.1
slowapi==0.1.9